# internal/transport/http/httpgen/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package httpgen

type APIV1MarketplaceParserServiceProductsSearchGetBadRequest ErrorResponse

func (*APIV1MarketplaceParserServiceProductsSearchGetBadRequest) aPIV1MarketplaceParserServiceProductsSearchGetRes() {
}

type APIV1MarketplaceParserServiceProductsSearchGetCode499 ErrorResponse

func (*APIV1MarketplaceParserServiceProductsSearchGetCode499) aPIV1MarketplaceParserServiceProductsSearchGetRes() {
}

type APIV1MarketplaceParserServiceProductsSearchGetGatewayTimeout ErrorResponse

func (*APIV1MarketplaceParserServiceProductsSearchGetGatewayTimeout) aPIV1MarketplaceParserServiceProductsSearchGetRes() {
}

type APIV1MarketplaceParserServiceProductsSearchGetInternalServerError ErrorResponse

func (*APIV1MarketplaceParserServiceProductsSearchGetInternalServerError) aPIV1MarketplaceParserServiceProductsSearchGetRes() {
}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Status  int    `json:"status"`
	Message string `json:"message"`
}

// GetStatus returns the value of Status.
func (s *ErrorResponse) GetStatus() int {
	return s.Status
}

// GetMessage returns the value of Message.
func (s *ErrorResponse) GetMessage() string {
	return s.Message
}

// SetStatus sets the value of Status.
func (s *ErrorResponse) SetStatus(val int) {
	s.Status = val
}

// SetMessage sets the value of Message.
func (s *ErrorResponse) SetMessage(val string) {
	s.Message = val
}

// NewOptFloat64 returns new OptFloat64 with value set to v.
func NewOptFloat64(v float64) OptFloat64 {
	return OptFloat64{
		Value: v,
		Set:   true,
	}
}

// OptFloat64 is optional float64.
type OptFloat64 struct {
	Value float64
	Set   bool
}

// IsSet returns true if OptFloat64 was set.
func (o OptFloat64) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptFloat64) Reset() {
	var v float64
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptFloat64) SetTo(v float64) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptFloat64) Get() (v float64, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptFloat64) Or(d float64) float64 {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/Product
type Product struct {
	Name         string  `json:"name"`
	Link         string  `json:"link"`
	Price        float64 `json:"price"`
	Rating       float64 `json:"rating"`
	ReviewsCount int     `json:"reviewsCount"`
}

// GetName returns the value of Name.
func (s *Product) GetName() string {
	return s.Name
}

// GetLink returns the value of Link.
func (s *Product) GetLink() string {
	return s.Link
}

// GetPrice returns the value of Price.
func (s *Product) GetPrice() float64 {
	return s.Price
}

// GetRating returns the value of Rating.
func (s *Product) GetRating() float64 {
	return s.Rating
}

// GetReviewsCount returns the value of ReviewsCount.
func (s *Product) GetReviewsCount() int {
	return s.ReviewsCount
}

// SetName sets the value of Name.
func (s *Product) SetName(val string) {
	s.Name = val
}

// SetLink sets the value of Link.
func (s *Product) SetLink(val string) {
	s.Link = val
}

// SetPrice sets the value of Price.
func (s *Product) SetPrice(val float64) {
	s.Price = val
}

// SetRating sets the value of Rating.
func (s *Product) SetRating(val float64) {
	s.Rating = val
}

// SetReviewsCount sets the value of ReviewsCount.
func (s *Product) SetReviewsCount(val int) {
	s.ReviewsCount = val
}

type SearchProductsResponse []Product

func (*SearchProductsResponse) aPIV1MarketplaceParserServiceProductsSearchGetRes() {}